+++ apps/aptos/components/Farms/components/FarmTable/Actions/ActionPanel.tsx
@@ -1,12 +1,14 @@
-import { useTranslation } from '@pancakeswap/localization'
-import { LinkExternal, Text, useMatchBreakpoints, Flex } from '@pancakeswap/uikit'
-import { FarmWidget } from '@pancakeswap/widgets-internal'
-import { styled, css, keyframes } from 'styled-components'
-import getLiquidityUrlPathParts from 'utils/getLiquidityUrlPathParts'
 import { FarmWithStakedValue } from '@pancakeswap/farms'
+import { useTranslation } from '@pancakeswap/localization'
+import { Flex, LinkExternal, Text, useMatchBreakpoints } from '@pancakeswap/uikit'
+import { FarmWidget } from '@pancakeswap/widgets-internal'
+import { getDisplayFarmCakePerSecond } from 'components/Farms/components/getDisplayFarmCakePerSecond'
+import { useFarms } from 'state/farms/hook'
+import { css, keyframes, styled } from 'styled-components'
+import getLiquidityUrlPathParts from 'utils/getLiquidityUrlPathParts'
 
 import Apr, { AprProps } from '../Apr'
-import { HarvestAction, HarvestActionContainer } from './HarvestAction'
+import { HarvestActionContainer, TableHarvestAction } from './HarvestAction'
 import StakedAction, { StakedContainer } from './StakedAction'
 
 const { Multiplier, Liquidity } = FarmWidget.FarmTable
@@ -20,6 +22,8 @@ export interface ActionPanelProps {
   expanded: boolean
   alignLinksToRight?: boolean
   isLastFarm: boolean
+  farmCakePerSecond?: string
+  totalMultipliers?: string
 }
 
 const expandAnimation = keyframes`
@@ -83,7 +87,7 @@ const ActionContainer = styled.div`
   display: flex;
   flex-direction: column;
 
-  ${({ theme }) => theme.mediaQueries.sm} {
+  ${({ theme }) => theme.mediaQueries.md} {
     flex-direction: row;
     align-items: center;
     flex-grow: 1;
@@ -117,7 +121,6 @@ const ActionPanel: React.FunctionComponent<React.PropsWithChildren<ActionPanelPr
 }) => {
   const farm = details
   const { isDesktop } = useMatchBreakpoints()
-
   const {
     t,
     currentLanguage: { locale },
@@ -129,6 +132,10 @@ const ActionPanel: React.FunctionComponent<React.PropsWithChildren<ActionPanelPr
     quoteTokenAddress: quoteToken?.address,
     tokenAddress: token?.address,
   })
+
+  const { totalRegularAllocPoint, cakePerBlock } = useFarms()
+  const totalMultipliers = totalRegularAllocPoint ? (Number(totalRegularAllocPoint) / 100).toString() : '0'
+  const farmCakePerSecond = getDisplayFarmCakePerSecond(farm.poolWeight?.toNumber(), cakePerBlock)
 
   return (
     <Container expanded={expanded} isLastFarm={isLastFarm}>
@@ -174,10 +181,23 @@ const ActionPanel: React.FunctionComponent<React.PropsWithChildren<ActionPanelPr
         )}
       </InfoContainer>
       <ActionContainer>
-        <HarvestActionContainer {...farm} userDataReady={userDataReady}>
-          {(props) => <HarvestAction {...props} />}
+        <HarvestActionContainer
+          pid={farm.pid}
+          lpAddress={farm.lpAddress}
+          earnings={farm?.userData?.earnings}
+          dual={farm.dual}
+          earningsDualTokenBalance={farm?.userData?.earningsDualTokenBalance}
+        >
+          {(props) => <TableHarvestAction {...props} />}
         </HarvestActionContainer>
-        <StakedContainer {...farm} userDataReady={userDataReady} lpLabel={lpLabel} displayApr={apr.value}>
+        <StakedContainer
+          {...farm}
+          userDataReady={userDataReady}
+          lpLabel={lpLabel}
+          displayApr={apr.value}
+          farmCakePerSecond={farmCakePerSecond}
+          totalMultipliers={totalMultipliers}
+        >
           {(props) => <StakedAction {...props} />}
         </StakedContainer>
       </ActionContainer>