+++ apps/aptos/components/Farms/components/FarmTable/Actions/HarvestAction.tsx
@@ -1,22 +1,22 @@
+import { Token } from '@pancakeswap/aptos-swap-sdk'
 import { TransactionResponse } from '@pancakeswap/awgmi/core'
-import { useTranslation } from '@pancakeswap/localization'
-import { Skeleton, useToast } from '@pancakeswap/uikit'
+import { useMatchBreakpoints } from '@pancakeswap/uikit'
 import { FarmWidget } from '@pancakeswap/widgets-internal'
-import { ToastDescriptionWithTx } from 'components/Toast'
-import useCatchTxError from 'hooks/useCatchTxError'
-import { BIG_ZERO } from '@pancakeswap/utils/bigNumber'
-import { getBalanceAmount } from '@pancakeswap/utils/formatBalance'
 import BigNumber from 'bignumber.js'
-import { usePriceCakeUsdc } from 'hooks/useStablePrice'
-import { FARM_DEFAULT_DECIMALS } from 'components/Farms/constants'
-import { FarmWithStakedValue } from '@pancakeswap/farms'
 import useHarvestFarm from '../../../hooks/useHarvestFarm'
+import HarvestAction from '../../FarmCard/HarvestAction'
 
-const { FarmTableHarvestAction } = FarmWidget.FarmTable
+const { ActionContainer } = FarmWidget.FarmTable
 
-interface HarvestActionProps extends FarmWithStakedValue {
-  userDataReady: boolean
+interface TableHarvestActionProps {
+  pid?: number
+  earnings?: BigNumber
+  dual?: {
+    token: Token
+  }
+  earningsDualTokenBalance?: BigNumber
   onReward: () => Promise<TransactionResponse>
+  onDone?: () => void
 }
 
 export const HarvestActionContainer = ({ children, ...props }) => {
@@ -25,48 +25,26 @@ export const HarvestActionContainer = ({ children, ...props }) => {
   return children({ ...props, onReward })
 }
 
-export const HarvestAction: React.FunctionComponent<React.PropsWithChildren<HarvestActionProps>> = ({
-  userData,
-  userDataReady,
+export const TableHarvestAction: React.FunctionComponent<React.PropsWithChildren<TableHarvestActionProps>> = ({
+  pid,
+  earnings,
+  dual,
+  earningsDualTokenBalance,
   onReward,
 }) => {
-  const { t } = useTranslation()
-  const { toastSuccess } = useToast()
-  const { fetchWithCatchTxError, loading: pendingTx } = useCatchTxError()
-  const earningsBigNumber = userData?.earnings ? new BigNumber(userData.earnings) : BIG_ZERO
-  const cakePrice = usePriceCakeUsdc()
-  let earnings = BIG_ZERO
-  let earningsBusd = 0
-  let displayBalance = userDataReady ? earnings.toFixed(5, BigNumber.ROUND_DOWN) : <Skeleton width={60} />
-
-  // If user didn't connect wallet default balance will be 0
-  if (!earningsBigNumber.isZero()) {
-    earnings = getBalanceAmount(earningsBigNumber, FARM_DEFAULT_DECIMALS)
-    earningsBusd = earnings.multipliedBy(cakePrice).toNumber()
-    displayBalance = earnings.toFixed(5, BigNumber.ROUND_DOWN)
-  }
-
-  const handleHarvest = async () => {
-    const receipt = await fetchWithCatchTxError(() => onReward())
-    if (receipt?.status) {
-      toastSuccess(
-        `${t('Harvested')}!`,
-        <ToastDescriptionWithTx txHash={receipt.transactionHash}>
-          {t('Your %symbol% earnings have been sent to your wallet!', { symbol: 'CAKE' })}
-        </ToastDescriptionWithTx>,
-      )
-    }
-  }
+  const { isDesktop } = useMatchBreakpoints()
 
   return (
-    <FarmTableHarvestAction
-      earnings={earnings}
-      earningsBusd={earningsBusd}
-      displayBalance={displayBalance}
-      pendingTx={pendingTx}
-      userDataReady={userDataReady}
-      handleHarvest={handleHarvest}
-    />
+    <ActionContainer style={{ minHeight: isDesktop ? 124.5 : 'auto' }}>
+      <HarvestAction
+        isTableView
+        pid={pid}
+        earnings={earnings}
+        dual={dual}
+        earningsDualTokenBalance={earningsDualTokenBalance}
+        onReward={onReward}
+      />
+    </ActionContainer>
   )
 }