+++ apps/aptos/components/Farms/components/FarmTable/Row.tsx
@@ -1,24 +1,25 @@
-import { useTranslation } from '@pancakeswap/localization'
 import { FarmWithStakedValue } from '@pancakeswap/farms'
+import { useDelayedUnmount } from '@pancakeswap/hooks'
+import { useTranslation } from '@pancakeswap/localization'
 import { Box, Flex, Skeleton, useMatchBreakpoints } from '@pancakeswap/uikit'
+import { BIG_ZERO } from '@pancakeswap/utils/bigNumber'
 import { FarmWidget } from '@pancakeswap/widgets-internal'
 import { createElement, useEffect, useRef, useState } from 'react'
 import { styled } from 'styled-components'
-import { BIG_ZERO } from '@pancakeswap/utils/bigNumber'
-import { useDelayedUnmount } from '@pancakeswap/hooks'
 
+import { EarnedUsdPrice } from '../FarmCard/EarnedUsdPrice'
 import ActionPanel from './Actions/ActionPanel'
 import Apr, { AprProps } from './Apr'
 import Farm from './Farm'
 
 const { MobileColumnSchema, DesktopColumnSchema } = FarmWidget
 const { FarmAuctionTag, CoreTag } = FarmWidget.Tags
-const { CellLayout, Details, Multiplier, Liquidity, Earned } = FarmWidget.FarmTable
+const { CellLayout, Details, Multiplier, Liquidity } = FarmWidget.FarmTable
 
 export interface RowProps {
   apr: AprProps
   farm: FarmWidget.FarmTableFarmTokenInfoProps
-  earned: FarmWidget.FarmTableEarnedProps
+  earned: FarmWithStakedValue
   multiplier: FarmWidget.FarmTableMultiplierProps
   liquidity: FarmWidget.FarmTableLiquidityProps
   details: FarmWithStakedValue
@@ -34,7 +35,7 @@ interface RowPropsWithLoading extends RowProps {
 const cells = {
   apr: Apr,
   farm: Farm,
-  earned: Earned,
+  earned: EarnedUsdPrice,
   details: Details,
   multiplier: Multiplier,
   liquidity: Liquidity,
@@ -184,14 +185,14 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
           </FarmMobileCell>
         </tr>
         <StyledTr onClick={toggleActionPanel}>
-          <td width="33%">
+          <td width="50%">
             <EarnedMobileCell>
               <CellLayout label={t('Earned')}>
-                <Earned {...props.earned} userDataReady={!!userDataReady} />
+                <EarnedUsdPrice {...props.earned} />
               </CellLayout>
             </EarnedMobileCell>
           </td>
-          <td width="33%">
+          <td>
             <AprMobileCell>
               <CellLayout label={t('APR')}>
                 <Apr
@@ -203,7 +204,7 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
               </CellLayout>
             </AprMobileCell>
           </td>
-          <td width="33%">
+          <td width="10%">
             <CellInner style={{ justifyContent: 'flex-end' }}>
               <Details actionPanelToggled={actionPanelExpanded} />
             </CellInner>