+++ apps/aptos/components/Farms/components/getDisplayApr.ts
@@ -1,9 +1,5 @@
-export const getDisplayApr = (cakeRewardsApr?: number, lpRewardsApr?: number) => {
-  if (cakeRewardsApr && lpRewardsApr) {
-    return (cakeRewardsApr + lpRewardsApr).toLocaleString('en-US', { maximumFractionDigits: 2 })
-  }
-  if (cakeRewardsApr) {
-    return cakeRewardsApr.toLocaleString('en-US', { maximumFractionDigits: 2 })
-  }
-  return null
+export const getDisplayApr = (cakeRewardsApr?: number, lpRewardsApr?: number, dualRewardsApr?: number) => {
+  return ((cakeRewardsApr ?? 0) + (lpRewardsApr ?? 0) + (dualRewardsApr ?? 0)).toLocaleString('en-US', {
+    maximumFractionDigits: 2,
+  })
 }