+++ apps/web/src/hooks/useSid.ts
@@ -1,9 +1,8 @@
 import { useMemo } from 'react'
-import useSWRImmutable from 'swr/immutable'
 import { namehash } from 'viem'
 import { getSidResolverContract } from 'utils/contractHelpers'
 import { useActiveChainId } from 'hooks/useActiveChainId'
-import { FetchStatus } from 'config/constants/types'
+import { useQuery } from '@tanstack/react-query'
 import { useSIDContract } from './useContract'
 
 function getSidAddress(networkId) {
@@ -29,8 +28,8 @@ export const useSidNameForAddress = (address: string, fetchData = true) => {
   const { chainId } = useActiveChainId()
   const sidContract = useSIDContract(getSidAddress(chainId), chainId)
 
-  const { data: sidName, status } = useSWRImmutable(
-    fetchData && address ? ['sidName', chainId, address.toLowerCase()] : null,
+  const { data: sidName, status } = useQuery(
+    ['sidName', chainId, address?.toLowerCase()],
     async () => {
       const reverseNode = `${address.toLowerCase().slice(2)}.addr.reverse`
       const reverseNameHash = namehash(reverseNode)
@@ -46,9 +45,15 @@ export const useSidNameForAddress = (address: string, fetchData = true) => {
         name: resolvedName,
       }
     },
+    {
+      enabled: Boolean(fetchData && address),
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      refetchOnMount: false,
+    },
   )
 
   return useMemo(() => {
-    return { sidName: sidName?.name, isLoading: status !== FetchStatus.Fetched }
+    return { sidName: sidName?.name, isLoading: status !== 'success' }
   }, [sidName, status])
 }