+++ apps/web/src/views/PositionManagers/hooks/useEarningTokenPriceInfo.ts
@@ -1,18 +1,18 @@
-import { useStablecoinPrice } from 'hooks/useBUSDPrice'
+import { Currency } from '@pancakeswap/sdk'
 import { getBalanceAmount } from '@pancakeswap/utils/formatBalance'
 import BigNumber from 'bignumber.js'
+import { useCurrencyUsdPrice } from 'hooks/useCurrencyUsdPrice'
 import { useMemo } from 'react'
-import { Currency } from '@pancakeswap/sdk'
 
 export const useEarningTokenPriceInfo = (earningToken: Currency, pendingReward: bigint | undefined) => {
-  const earningTokenPrice = useStablecoinPrice(earningToken ?? undefined, { enabled: !!earningToken })
+  const { data: earningTokenPrice } = useCurrencyUsdPrice(earningToken ?? undefined)
   const earningsBalance = useMemo(
     () => getBalanceAmount(new BigNumber(pendingReward?.toString() ?? 0), earningToken.decimals).toNumber(),
     [pendingReward, earningToken],
   )
 
   const earningUsdValue = useMemo(
-    () => new BigNumber(earningsBalance).times(earningTokenPrice?.toSignificant() ?? 0).toNumber(),
+    () => new BigNumber(earningsBalance).times(earningTokenPrice ?? 0).toNumber(),
     [earningsBalance, earningTokenPrice],
   )
   return { earningUsdValue, earningTokenPrice, earningsBalance }