+++ packages/smart-router/evm/constants/stableSwap/pools.ts
@@ -1,8 +1,9 @@
 import { ChainId } from '@pancakeswap/chains'
 
 import { StableSwapPool } from './types'
-import { pools as bscPools } from './56'
-import { pools as bscTestnetPools } from './97'
+import { pools as bscPools } from './bsc'
+import { pools as arbPools } from './arb'
+import { pools as bscTestnetPools } from './bscTestnet'
 
 export type StableSwapPoolMap<TChainId extends number> = {
   [chainId in TChainId]: StableSwapPool[]
@@ -15,11 +16,12 @@ export const isStableSwapSupported = (chainId: number | undefined): chainId is S
   return STABLE_SUPPORTED_CHAIN_IDS.includes(chainId)
 }
 
-export const STABLE_SUPPORTED_CHAIN_IDS = [ChainId.BSC, ChainId.BSC_TESTNET] as const
+export const STABLE_SUPPORTED_CHAIN_IDS = [ChainId.BSC, ChainId.BSC_TESTNET, ChainId.ARBITRUM_ONE] as const
 
 export type StableSupportedChainId = (typeof STABLE_SUPPORTED_CHAIN_IDS)[number]
 
 export const STABLE_POOL_MAP = {
   [ChainId.BSC]: bscPools,
   [ChainId.BSC_TESTNET]: bscTestnetPools,
+  [ChainId.ARBITRUM_ONE]: arbPools,
 } satisfies StableSwapPoolMap<StableSupportedChainId>